refactor(dispatch): extract lite adapters, team widget, and portal dropdown
